Create quiver3MagColour.py
Quiver3 doesn't natively support colured arrows according to vector magnitude so I wrote this script to plot wind velocity measurements for my Mechanical Engineering final year project. I had a ton of help tweaking quiver from Greg, who I'll add as a collaborator when I've the chance.

We intend to make this more automated but anyone that's used Python3 and has the required repositories should be able to manipulate this for their needs.